Allow the outsideClickDeselects option to be a function
This relates to #401. I have some external elements where I need to
click and type, while I need to keep outsideClickDeselects set to false
for other reasons. So I changed outsideClickDeselects to allow being a
function that takes the click target as a param and returns a boolean.
This lets the caller decide whether to deselect or not.